modeler/requirements.txt
Flask
coverage
pytest
google-cloud-bigquery